Split into a library and a binary
Some basic refactoring to split unciv into a library, which can be used
by other code, and a binary which relies on it.

The code for writing files (including the png files for RIM images) is now
a part of the binary, the libary technically doesn't depend on png (though
Cargo.toml hasn't been updated to mention this).

The API is not perfect, but the binary should give a basic example on how
to use it. ZfsEntry structs refer to an entry before it is read: you need
to provide a reader to the same file to get its data. On the other hand,
RimImage _does_ contain the data, once you load it from a ZfsEntry, there's
no need to keep the file/reader around.