Make the png dependency optional.
Currently, we use the 'png' library to convert 'rim' images to png files.
However, this is constrained to the unciv program, and the library has no need
of the 'png' dependency.

Make the 'png' dependency optional so that users of the library don't gain a
new, heavy dependency. It's behind the 'png' feature, which is currently
disabled by default (though this may change… I'm still thinking about it!)

Note that 'png' is still at version 0.15 (to match the original's support of
Debian bullseye). Additionally, note that the 'png' dependency is not an
implicit dependency (though that might be better) as they're not supported by
cargo 1.46. That, too, may change at some point.

If 'png' is disabled, unciv will extract .rim files as they are.